* snd_codec.c (S_CodecUtilOpen):  cast for length must be long.
git-svn-id: file:///svn/uhexen2/code/trunk@3896 eaf649e6-6c13-0410-b194-84221022858e

+++ engine/h2shared/snd_codec.c
@@ -253,7 +253,7 @@ snd_stream_t *S_CodecUtilOpen(const char *filename, snd_codec_t *codec)
 	stream->fh.file = handle;
 	stream->fh.start = ftell(handle);
 	stream->fh.pos = 0;
-	stream->fh.length = (int) length;
+	stream->fh.length = (long)length;
 	stream->fh.pak = stream->pak = pak;
 	return stream;
 }
